Creation of broken obstacle file and changes to movement script
Obstacle script is not being used. was suppose to be used to control player movement, however did not work as player movement was affected in the process. therefore i have added the player death inside the playermovement script. I have also made the serielzed obstacle an array so that we can add all obstavles to kill the player