Creation of death screen
When player collides with obstacle, death screen is displayed and restart button appears, restart button also lightens when player hovers over. When restart button is pressed game is restarted. 1 issue that needs to be resolved in a later stage is: bottle count does not reset to 0. So game will continue to count bottles from previous run.